Move theano.compile.ops tests into their test module
